MAINT: move _util to _detail
