+++ torch_np/__init__.py
@@ -6,7 +6,7 @@ from ._wrapper import *
 from ._unary_ufuncs import *
 from ._binary_ufuncs import *
 from ._ndarray import can_cast, result_type, newaxis
-from ._util import AxisError, UFuncTypeError
+from ._detail._util import AxisError, UFuncTypeError
 from ._getlimits import iinfo, finfo
 from ._getlimits import errstate
